feat: 5단계 JWT
- Access Token 만료시간 60분
- 토큰이 필요한 API 요청에서 정상 토큰이 아닐 때
  - 에러 메세지: 토큰이 유효하지 않습니다., StatusCode: 401
- 유효한 토큰이지만 해당 사용자가 작성한 댓글이 아닐 때
  -  에러 메세지: 작성자만 삭제/수정할 수 있습니다., StatusCode: 403
- DB에 이미 존재하는 username으로 회원가입을 요청할 때, 요구조건으로 에러 메세지 수정
  - 에러 메세지: 중복된 username 입니다.
- 로그인 시, 전달된 username과 password 중 맞지 않는 정보가 있을 때, 요구조건으로 에러 메세지 수정
  - 에러 메시지: 회원을 찾을 수 없습니다.
- StatusCode 나누기
  - 처음 공통 조건인 400으로 통일이 아닌 401, 403 사용
- 1~4 단계에서 인증/인가가 완료된 후에만 기능이 동작하도록 수정한다.

+++ src/main/java/com/sparta/assignment01/service/CommentService.java
@@ -4,14 +4,21 @@ import com.sparta.assignment01.dto.CommentRequestDto;
 import com.sparta.assignment01.dto.CommentResponseDto;
 import com.sparta.assignment01.entity.Comment;
 import com.sparta.assignment01.entity.Schedule;
+import com.sparta.assignment01.entity.User;
+import com.sparta.assignment01.exception.AccessDeniedException;
 import com.sparta.assignment01.exception.NotFoundException;
+import com.sparta.assignment01.exception.UnauthorizedException;
+import com.sparta.assignment01.jwt.JwtUtil;
 import com.sparta.assignment01.repo.CommentRepo;
 import com.sparta.assignment01.repo.ScheduleRepo;
+import com.sparta.assignment01.repo.UserRepo;
+import io.jsonwebtoken.Claims;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.StringUtils;
 
 @Service
 @RequiredArgsConstructor
@@ -19,11 +26,17 @@ public class CommentService {
 
     private final CommentRepo commentRepo;
     private final ScheduleRepo scheduleRepo;
+    private final UserRepo userRepo;
+    private final JwtUtil jwtUtil;
 
-    public ResponseEntity<CommentResponseDto> createComment(Long scheduleId, CommentRequestDto commentRequestDto) {
+    public ResponseEntity<CommentResponseDto> createComment(Long scheduleId, CommentRequestDto commentRequestDto, String bearerToken) {
+
+        String jwtToken = checkToken(bearerToken);
+        Claims info = jwtUtil.getUserInfoFromToken(jwtToken);
 
         Schedule schedule = findScheduleById(scheduleId);
-        Comment comment = new Comment(commentRequestDto, schedule);
+        User user = findUserByName(info.getSubject());
+        Comment comment = new Comment(commentRequestDto, schedule, user);
 
         commentRepo.save(comment);
 
@@ -31,18 +44,34 @@ public class CommentService {
     }
 
     @Transactional
-    public ResponseEntity<CommentResponseDto> updateComment(Long commentId, CommentRequestDto commentRequestDto) {
+    public ResponseEntity<CommentResponseDto> updateComment(Long commentId, CommentRequestDto commentRequestDto, String bearerToken) {
+
+        String jwtToken = checkToken(bearerToken);
+        Claims info = jwtUtil.getUserInfoFromToken(jwtToken);
 
         Comment comment = findCommentById(commentId);
+        User user = findUserByName(info.getSubject());
+
+        if (!comment.getUser().getId().equals(user.getId())) {
+            throw new AccessDeniedException("작성자만 수정할 수 있습니다.");
+        }
 
         comment.update(commentRequestDto);
 
         return new ResponseEntity<>(new CommentResponseDto(comment), HttpStatus.OK);
     }
 
-    public ResponseEntity<String> deleteComment(Long commentId) {
+    public ResponseEntity<String> deleteComment(Long commentId, String bearerToken) {
+
+        String jwtToken = checkToken(bearerToken);
+        Claims info = jwtUtil.getUserInfoFromToken(jwtToken);
 
         Comment comment = findCommentById(commentId);
+        User user = findUserByName(info.getSubject());
+
+        if (!comment.getUser().getId().equals(user.getId())) {
+            throw new AccessDeniedException("작성자만 삭제할 수 있습니다.");
+        }
 
         commentRepo.delete(comment);
 
@@ -55,6 +84,21 @@ public class CommentService {
 
     public Comment findCommentById(Long id) {
         return commentRepo.findById(id).orElseThrow( () -> new NotFoundException("해당 댓글은 존재하지 않습니다."));
+    }
+
+    public User findUserByName(String name) {
+        return userRepo.findByName(name).orElseThrow( () -> new NotFoundException("해당 회원은 존재하지 않습니다."));
+    }
+
+    public String checkToken(String bearerToken) {
+
+        String jwtToken = jwtUtil.getJwtFromHeader(bearerToken);
+
+        if (!StringUtils.hasText(jwtToken) || !jwtUtil.validateToken(jwtToken)) {
+            throw new UnauthorizedException("토큰이 유효하지 않습니다.");
+        }
+
+        return jwtToken;
     }
 
 }
